Setting Maven user settings as a part of IBeforeTest can be disabled and configured
