+++ tests/org.jboss.tools.maven.ui.bot.test/build.properties
@@ -1,5 +1,4 @@
 source.. = src/
 output.. = bin/
 bin.includes = META-INF/,\
-               .,\
-               plugin.xml
+               .
