+++ tests/org.jboss.tools.usercase.ticketmonster.ui.bot.test/src/org/jboss/tools/usercase/ticketmonster/ui/bot/test/TicketMonsterBaseTest.java
@@ -111,7 +111,8 @@ public class TicketMonsterBaseTest {
 		MavenUserPreferencePage mpreferences = new MavenUserPreferencePage();
 		preferenceDialog.select(mpreferences);
 		mpreferences.setUserSettings(new File(USER_SETTINGS).getAbsolutePath());
-		mpreferences.ok();
+		mpreferences.apply();
+		preferenceDialog.ok();
 	}
 	
 	public void createTicketMonsterEAP6(){
